update c++ to latest upstream

+++ core-cpp/swa/src/Duration.cc
@@ -1,6 +1,25 @@
-//
-// UK Crown Copyright (c) 2016. All Rights Reserved.
-//
+/*
+ * ----------------------------------------------------------------------------
+ * (c) 2005-2023 - CROWN OWNED COPYRIGHT. All rights reserved.
+ * The copyright of this Software is vested in the Crown
+ * and the Software is the property of the Crown.
+ * ----------------------------------------------------------------------------
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * ----------------------------------------------------------------------------
+ * Classification: UK OFFICIAL
+ * ----------------------------------------------------------------------------
+ */
+
 #include "swa/Duration.hh"
 #include "swa/String.hh"
 #include "swa/ProgramError.hh"
@@ -595,7 +614,7 @@ namespace SWA
     // PT1M0.01S   2       no      yes       PT1M0.01S 
     // PT1M0.001S  2       no      yes       PT1M0.00S NB does not hide trailing zero field if truncate false
 
-    long double value = static_cast<long double>(nanosRemaining) / divisors[smallestField];
+    long double value = static_cast<long double>(nanosRemaining) / static_cast<int64_t>(divisors[smallestField]);
     if ( value > 0.0 || !truncateDecimal || !hideZeros || result == prefix.s_str() )
     {
       addTimePrefix = addTimePrefix || smallestField == Hour;
